requirements.txt
python-dotenv
openai
qdrant-client
aiohttp
tiktoken
